📦️ Chore: 주석 처리 및 예외 처리 추가
