+++ src/main/java/com/likelion/backendplus4/talkpick/batch/news/article/infrastructure/collector/config/batch/StepConfig.java
@@ -15,7 +15,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.task.TaskExecutor;
 import org.springframework.transaction.PlatformTransactionManager;
 
-import com.likelion.backendplus4.talkpick.batch.news.article.application.exception.ArticleCollectorException;
+import com.likelion.backendplus4.talkpick.batch.news.article.exception.ArticleCollectorException;
 import com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.jpa.entity.ArticleEntity;
 
 @Configuration
